pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Sara <ssalvador@kiteris.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"  # Añade <4.0 para compatibilidad con langchain-community
# Las dependencias se añadirán automáticamente con `poetry add`
langchain = "^0.3.21"
langchain-openai = "^0.3.9"
langchain-community = "^0.3.20"
python-dotenv = "^1.0.1"
qdrant-client = "^1.13.3"
sentence-transformers = "^3.4.1"
pypdf = "^5.4.0"
streamlit = "^1.43.2"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
langchain-huggingface = "^0.1.2"
langchain-qdrant = "^0.2.0"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
